1. 통계에서 통계 지표 데이터 노출 (sum, min, max, avg, count)

+++ src/main/java/com/project/nameMaker/stats/entity/NameStats.java
@@ -34,7 +34,7 @@ public class NameStats {
     private Integer totalRank;
 
     @NotEmpty
-    private Integer yearCount;
+    private Long yearCount;
 
     /*
     이름
